@@ -4,52 +4,51 @@ using McMaster.Extensions.CommandLineUtils;
 using DevLab.JmesPath;
 using System.Text.Json;
 
-namespace Hal9000Cli
+namespace Hal9000Cli;
+
+[Command(Name = "list", Description = "Get list of todo items")]
+[HelpOption]
+public class QueryTodoItemCmd
 {
-    [Command(Name = "list", Description = "Get list of todo items")]
-    [HelpOption]
-    public class QueryTodoItemCmd
+    private readonly IConsole _console;
+    private readonly TodoService _todoService;
+
+    public QueryTodoItemCmd(IConsole console, TodoService todoService)
     {
-        private readonly IConsole _console;
-        private readonly TodoService _todoService;
+        _console = console;
+        _todoService = todoService;
+    }
 
-        public QueryTodoItemCmd(IConsole console, TodoService todoService)
+    [Option(CommandOptionType.SingleValue, ShortName = "", LongName = "query", Description = " JMESPath query string. See http://jmespath.org/ for more information and examples.", ShowInHelpText = true, ValueName = "")]
+    public string? Query { get; set; }
+
+    protected Task<int> OnExecute(CommandLineApplication app)
+    {
+        var items = _todoService.GetItems();
+
+        var options = new JsonSerializerOptions { WriteIndented = true };
+        string jsonString = JsonSerializer.Serialize(items, options);
+
+        if (!string.IsNullOrEmpty(Query))
         {
-            _console = console;
-            _todoService = todoService;
+            jsonString = JMESPathTransform(jsonString);
+            jsonString = JsonSerializer.Serialize(JsonSerializer.Deserialize<object>(jsonString), options);
         }
 
-        [Option(CommandOptionType.SingleValue, ShortName = "", LongName = "query", Description = " JMESPath query string. See http://jmespath.org/ for more information and examples.", ShowInHelpText = true, ValueName = "")]
-        public string? Query { get; set; }
+        _console.WriteLine(jsonString);
 
-        protected Task<int> OnExecute(CommandLineApplication app)
+        return Task.FromResult(0);
+    }
+
+    private string JMESPathTransform(string input)
+    {
+        if (!string.IsNullOrEmpty(Query))
         {
-            var items = _todoService.GetItems();
-
-            var options = new JsonSerializerOptions { WriteIndented = true };
-            string jsonString = JsonSerializer.Serialize(items, options);
-
-            if (!string.IsNullOrEmpty(Query))
-            {
-                jsonString = JMESPathTransform(jsonString);
-                jsonString = JsonSerializer.Serialize(JsonSerializer.Deserialize<object>(jsonString), options);
-            }
-
-            _console.WriteLine(jsonString);
-
-            return Task.FromResult(0);
+            var jmes = new JmesPath();
+            return jmes.Transform(input, Query);
         }
 
-        private string JMESPathTransform(string input)
-        {
-            if (!string.IsNullOrEmpty(Query))
-            {
-                var jmes = new JmesPath();
-                return jmes.Transform(input, Query);
-            }
-
-            return string.Empty;
-        }
+        return string.Empty;
     }
 }
 
